# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth": {
            "post": {
                "description": "Register",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":1006, \"msg\": \"添加账户失败\", \"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/delete": {
            "delete": {
                "description": "DeleteSpace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":1006, \"msg\": \"删除失败\", \"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get": {
            "get": {
                "description": "GetSpace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":1006, \"msg\": \"获取失败\", \"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getlist": {
            "get": {
                "description": "GetList",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":1006, \"msg\": \"获取失败\", \"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/test": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{ \"name\": \"1\",\"password\":\"2\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updavatar": {
            "put": {
                "description": "UpdAvatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "{\"code\":1006, \"msg\": \"更新失败\", \"data\": \"data\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}